refactor: deduplicate formatTime into shared formatUtil
Amp-Thread-ID: https://ampcode.com/threads/T-019cfee8-bad9-7131-8e7c-d2444c3fc456
