+++ src/composables/useWaveAudioPlayer.ts
@@ -3,6 +3,7 @@ import { computed, onUnmounted, ref, watch } from 'vue'
 import type { Ref } from 'vue'
 
 import { api } from '@/scripts/api'
+import { formatTime } from '@/utils/formatUtil'
 
 interface WaveformBar {
   height: number
@@ -183,11 +184,4 @@ export function useWaveAudioPlayer(options: UseWaveAudioPlayerOptions) {
     seekToRatio,
     handleWaveformClick
   }
-}
-
-function formatTime(seconds: number): string {
-  if (isNaN(seconds) || seconds === 0) return '0:00'
-  const mins = Math.floor(seconds / 60)
-  const secs = Math.floor(seconds % 60)
-  return `${mins}:${secs.toString().padStart(2, '0')}`
 }
