+++ src/renderer/extensions/vueNodes/widgets/utils/audioUtils.ts
@@ -1,16 +1,7 @@
 import type { ResultItemType } from '@/schemas/apiSchema'
 import { app } from '@/scripts/app'
 
-/**
- * Format time in MM:SS format
- */
-export function formatTime(seconds: number): string {
-  if (isNaN(seconds) || seconds === 0) return '0:00'
-
-  const mins = Math.floor(seconds / 60)
-  const secs = Math.floor(seconds % 60)
-  return `${mins}:${secs.toString().padStart(2, '0')}`
-}
+export { formatTime } from '@/utils/formatUtil'
 
 export function getResourceURL(
   subfolder: string,
